move test to connection class
