+++ csharp/test/Drivers/Databricks/StatementTests.cs
@@ -486,54 +486,5 @@ namespace Apache.Arrow.Adbc.Tests.Drivers.Databricks
             Assert.True(batchCount > 1, "Should have read multiple batches");
             OutputHelper?.WriteLine($"Successfully read {totalRows} rows in {batchCount} batches after 30 minute delay with {configName}");
         }
-
-        [SkippableFact]
-        public async Task SetDefaultCatalogAndSchemaOptionsTest()
-        {
-            string? defaultCatalog = null;
-            string? defaultSchema = null;
-            if (!string.IsNullOrEmpty(TestConfiguration.Catalog))
-            {
-                defaultCatalog = TestConfiguration.Catalog;
-            }
-            else
-            {
-                Skip.If(true, "No catalog specified in environment variable DATABRICKS_CATALOG or test configuration");
-            }
-
-            if (!string.IsNullOrEmpty(TestConfiguration.DbSchema))
-            {
-                defaultSchema = TestConfiguration.DbSchema;
-            }
-
-            AdbcConnection connection = NewConnection();
-            AdbcStatement statement = connection.CreateStatement();
-
-            // Verify the settings were applied by querying the current catalog and schema
-            statement.SqlQuery = "SELECT current_catalog()" + (defaultSchema != null ? ", current_schema()" : "");
-            var result = await statement.ExecuteQueryAsync();
-            Assert.NotNull(result.Stream);
-
-            var batch = await result.Stream.ReadNextRecordBatchAsync();
-            Assert.NotNull(batch);
-            Assert.Equal(1, batch.Length);
-
-            // Get the values from the result
-            var catalogArray = (StringArray)batch.Column(0);
-
-            string actualCatalog = catalogArray.GetString(0) ?? string.Empty;
-
-            // Verify the values match what we set
-            Assert.Equal(defaultCatalog, actualCatalog);
-
-            if (defaultSchema != null)
-            {
-                var schemaArray = (StringArray)batch.Column(1);
-                string actualSchema = schemaArray.GetString(0) ?? string.Empty;
-                Assert.Equal(defaultSchema, actualSchema);
-            }
-
-            OutputHelper?.WriteLine($"Successfully set and verified default catalog: {defaultCatalog} and schema: {defaultSchema}");
-        }
     }
 }
